+++ webpack/extra-webpack.config.js
@@ -1,6 +1,131 @@
 const path = require('path');
+const fs = require('fs');
+const readXlsxFile = require('read-excel-file/node');
+const JSON5 = require('json5');
+const json5Writer = require('json5-writer');
 
-module.exports = (config) => {
+const xlsxFolder = './src/conf/xlsx';
+const csvFolder = './src/conf/csv';
+const settingsPath = './src/assets/settings.json5';
+const definitionsFilePropName = 'definitionsFile';
+
+/**
+ * Converts array of cell values to CSV row
+ * @param {Array} row
+ * @return {string}
+ */
+function createCsvRow(row) {
+  return row
+    .map((cell) => {
+      if (/["\s,]/.test(cell)) {
+        return '"' + cell.replace(/"/g, '""') + '"';
+      } else {
+        return cell;
+      }
+    })
+    .join(',');
+}
+
+/**
+ * Returns an object which maps service base URLs to CSV data extracted from XLSX files.
+ * @returns {Promise<Object>}
+ */
+async function prepareCsvData() {
+  // Cell value to mark the beginning of data
+  const marker = '---SERVICE BASE URL:';
+  // Index offset of the row from which the data starts relative to the row with URL
+  const dataOffset = 3;
+  // Read all XLSX files in the folder
+  const fileNames = fs.readdirSync(xlsxFolder);
+  const url2desc = {};
+  for (let i = 0; i < fileNames.length; i++) {
+    if (/(.*)\.xlsx$/.test(fileNames[i])) {
+      const filename = RegExp.$1;
+      const fromFilename = filename + '.xlsx';
+      const sheets = await readXlsxFile(xlsxFolder + '/' + fromFilename, {
+        getSheets: true
+      });
+
+      for (let j = 1; j <= sheets.length; j++) {
+        const rows = await readXlsxFile(xlsxFolder + '/' + fromFilename, {
+          sheet: j
+        });
+
+        const urlIndex = rows.findIndex((row) => row[0] === marker);
+        const urls = rows[urlIndex][1].split(',');
+        const desc = rows.slice(urlIndex + dataOffset).filter(
+          ([, , type, , hideShow, , ,]) =>
+            // Skip hidden search parameters
+            type !== 'search parameter' || hideShow !== 'hide'
+        );
+        urls.forEach((url) => {
+          url2desc[url] = (url2desc[url] || []).concat(desc);
+        });
+      }
+    }
+  }
+  return url2desc;
+}
+
+/**
+ * Updates settings.json5 and creates CSV-files.
+ * @param {Object} url2desc - object which maps service base URLs to CSV
+ *   data extracted from XLSX files.
+ */
+function updateAppSettings(url2desc) {
+  const allUrls = Object.keys(url2desc);
+  const updateSettingsObj = { default: {}, customization: {} };
+  const settings = JSON5.parse(fs.readFileSync(settingsPath).toString());
+
+  // Remove references to previous CSV data files in the settings before placing new references
+  delete updateSettingsObj.default[definitionsFilePropName];
+  if (updateSettingsObj.customization) {
+    Object.keys(updateSettingsObj.customization).forEach((url) => {
+      if (updateSettingsObj.customization[url]) {
+        delete updateSettingsObj.customization[url][definitionsFilePropName];
+      }
+    });
+  }
+
+  for (let i = 0; i < allUrls.length; i++) {
+    const url = allUrls[i];
+    const toFilename = 'desc-' + i + '.csv';
+    if (url === 'default') {
+      Object.keys(settings.default).forEach((key) => {
+        updateSettingsObj.default[key] = undefined;
+      });
+      updateSettingsObj.default[definitionsFilePropName] = toFilename;
+    } else {
+      if (!updateSettingsObj.customization) {
+        updateSettingsObj.customization = {};
+      }
+      updateSettingsObj.customization[url] = {};
+      if (settings.customization[url]) {
+        Object.keys(settings.customization[url]).forEach((key) => {
+          updateSettingsObj.customization[url][key] = undefined;
+        });
+      }
+      updateSettingsObj.customization[url][
+        definitionsFilePropName
+      ] = toFilename;
+    }
+
+    fs.writeFileSync(
+      csvFolder + '/' + toFilename,
+      url2desc[url].map((row) => createCsvRow(row)).join('\n')
+    );
+  }
+
+  const settingsWriter = json5Writer.load(
+    fs.readFileSync(settingsPath).toString()
+  );
+  settingsWriter.write(updateSettingsObj);
+  fs.writeFileSync(settingsPath, settingsWriter.toSource());
+}
+
+module.exports = async (config) => {
+  updateAppSettings(await prepareCsvData());
+
   config.module.rules.push(
     {
       test: /definitions\/index.json$/,